Return page size as unsigned long for riscv64
Link: https://github.com/checkpoint-restore/criu/commit/28adebefb759306959df41acbe9cb74bfebd03ea

+++ include/common/arch/riscv64/asm/page.h
@@ -37,7 +37,7 @@ static inline unsigned page_shift(void)
 
 #else /* CR_NOGLIBC */
 
-extern unsigned page_size(void);
+extern unsigned long page_size(void);
 #define PAGE_SIZE page_size()
 
 #endif /* CR_NOGLIBC */
